requirements.txt
crewai>=0.28.8
crewai_tools>=0.1.6
langchain_community>=0.0.29
langchain>=0.1.13
langchain-openai>=0.1.7
langchain-cohere>=0.1.5
instructor>=0.5.2,<0.6.0
opentelemetry-sdk>=1.22.0,<2.0.0
openai>=1.13.3,<2.0.0
python-dotenv==1.0.0
click>=8.1.7,<9.0.0
opentelemetry-exporter-otlp-proto-http>=1.22.0,<2.0.0
appdirs>=1.4.4,<2.0.0
embedchain>=0.1.98,<0.2.0
opentelemetry-api>=1.22.0,<2.0.0
regex>=2023.12.25,<2024.0.0
pydantic>=2.4.2,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
pyright>=1.1.350,<2.0.0
lancedb>=0.5.4,<0.6.0
requests>=2.31.0,<3.0.0
chromadb>=0.4.22,<0.5.0
selenium>=4.18.1,<5.0.0
pytube>=15.0.0,<16.0.0
pytest>=8.0.0,<9.0.0
PyYAML>=5.3
numpy>=1,<2
tenacity>=8.1.0,<9.0.0
dataclasses-json>=0.5.7,<0.7
SQLAlchemy>=1.4,<3
langchain-core>=0.1.33,<0.2.0
langsmith>=0.1.0,<0.2.0
aiohttp>=3.8.3,<4.0.0